requirements.txt
flask
flask-cors
gunicorn
numpy
opencv-python 
pillow 
pytesseract